Corrección rutas.
