+++ src/app/(loggedin)/teacher/add/TeacherRegistrationFormDialog.tsx
@@ -4,7 +4,7 @@ import {Label} from "@/components/ui/label";
 import {Select, SelectContent, SelectItem, SelectTrigger, SelectValue} from "@/components/ui/select";
 import {X} from "lucide-react";
 import {useState} from "react";
-import {GradeWithSubjects} from "@/app/(loggedin)/parent/add/page";
+import {GradeWithSubjects} from "@/app/(loggedin)/teacher/add/page";
